+++ packages/codenic_bloc_use_case/lib/src/util/either_extension.dart
@@ -2,9 +2,24 @@ import 'package:fpdart/fpdart.dart';
 
 /// Extension for [Either] to get the value of [Right] or [Left].
 extension EitherExtension<L, R> on Either<L, R> {
+  /// Returns the value of [Right] if available. Otherwise, a [StateError]
+  /// is thrown.
+
+  R right() => fold(
+        (l) => throw StateError('The value is not a right value: $l'),
+        (r) => r,
+      );
+
+  /// Returns the value of [Left] if available. Otherwise, a [StateError]
+  /// is thrown.
+  L left() => fold(
+        (l) => l,
+        (r) => throw StateError('The value is not a left value: $r'),
+      );
+
   /// Returns the value of [Right] if available. Otherwise, `null` is returned.
-  R? getRightOrNull() => fold((l) => null, (r) => r);
+  R? rightOrNull() => fold((l) => null, (r) => r);
 
   /// Returns the value of [Left] if available. Otherwise, `null` is returned.
-  L? getLeftOrNull() => fold((l) => l, (r) => null);
+  L? leftOrNull() => fold((l) => l, (r) => null);
 }